Move token cache to internal package
